+++ src/main/java/com/github/zighang_zighang/global/config/SecurityConfig.java
@@ -2,6 +2,8 @@ package com.github.zighang_zighang.global.config;
 
 import com.github.zighang_zighang.global.auth.filter.JwtAuthenticationFilter;
 import com.github.zighang_zighang.global.auth.service.CustomOAuth2UserService;
+import com.github.zighang_zighang.global.auth.handler.OAuth2AuthenticationSuccessHandler;
+import com.github.zighang_zighang.global.auth.handler.OAuth2AuthenticationFailureHandler;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,6 +22,8 @@ public class SecurityConfig {
     private final JwtAuthenticationFilter jwtAuthenticationFilter;
     private final CorsConfigurationSource corsConfigurationSource;
     private final CustomOAuth2UserService customOAuth2UserService;
+    private final OAuth2AuthenticationSuccessHandler oAuth2AuthenticationSuccessHandler;
+    private final OAuth2AuthenticationFailureHandler oAuth2AuthenticationFailureHandler;
 
     @Bean
     public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
@@ -31,7 +35,6 @@ public class SecurityConfig {
                 .anyRequest().authenticated()
             )
             .oauth2Login(oauth2 -> oauth2
-                .defaultSuccessUrl("/auth/oauth2-success", true)
                 .loginPage("/login")
                 .redirectionEndpoint(redirection -> redirection
                     .baseUri("/login/oauth2/code/*")
@@ -39,12 +42,8 @@ public class SecurityConfig {
                 .userInfoEndpoint(userInfo -> userInfo
                     .userService(customOAuth2UserService)
                 )
-                .failureUrl("/login?error=oauth2")
-                .failureHandler((request, response, exception) -> {
-                    System.err.println("OAuth2 로그인 실패: " + exception.getMessage());
-                    exception.printStackTrace();
-                    response.sendRedirect("/login?error=oauth2&details=" + exception.getClass().getSimpleName());
-                })
+                .successHandler(oAuth2AuthenticationSuccessHandler)
+                .failureHandler(oAuth2AuthenticationFailureHandler)
             )
             .addFilterBefore(jwtAuthenticationFilter, UsernamePasswordAuthenticationFilter.class);
